GlobalsView: rename file, use DeepEqual for updates
